Remove unneeded org.json dependency
This gets rid of warnings about multiple occurrences of org.json.JSONObject
These occur because it is also supplied by the com.vaadin.external.google:
android-json library brought in by the spring-boot-starter-test.
The latter only has test scope so would exclude that if needed this for
the working application but it seems to work without the explicit org.json
dependency so the running application presumably doesn't need it.